Removed email related things
