+++ src/main/resources/application-local.yml
@@ -15,17 +15,7 @@ spring:
         show-sql: true
         use_sql_comments: true
     database-platform: org.hibernate.dialect.PostgreSQLDialect
-  mail:
-    host: smtp.gmail.com
-    port: 587
-    username: your_email
-    password: your_app_password
-    properties:
-      mail:
-        smtp:
-          auth: true
-          starttls:
-            enable: true
+
 verify:
   base:
     frontend: http://localhost:3000/
