+++ src/main/resources/application.yml
@@ -34,18 +34,6 @@ spring:
         show-sql: true
         use_sql_comments: true
     database-platform: org.hibernate.dialect.PostgreSQLDialect
-  mail:
-    host: smtp.gmail.com
-    port: 587
-    username: ${SMTP_EMAIL}
-    password: ${SMTP_PASSWORD}
-    properties:
-      mail:
-        smtp:
-          auth: true
-          starttls:
-            enable: true
-            required: true
 
 jwt:
   secret: nyanpasu
